Move files to go/migration and rename
